Adding a function, approxD, to implement derivative approximation for any specified legitimate R function.  Also started writing a function to compute the intersection points of the tangent lines for the upper envelope.
Created a file for tests and wrote some for approxD.  I included complete roxygen descriptions for the functions even though the project only specifies we need one for the main function, just so we have something to go off of later.  Also if we want to host this on our own github pages later where people can download it then it might be helpful to have them anyway.